requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
dgl
gunicorn==20.1.0
idna==3.3
joblib==1.1.0
Markdown==3.3.4
networkx==2.6.3
numpy==1.21.2
Pillow==8.4.0
python-dateutil==2.8.2
pytz==2021.3
rdkit-pypi==2021.3.5.1
requests==2.26.0
scipy==1.7.1
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.0.0
uvicorn

httptools
httpx
aiofiles==0.6.0
fastapi==0.64.0
Jinja2==2.11.3
pytest==6.2.4

celery==5.1.2
redis==3.5.3
flower

torch==1.9.1+cpu
typing-extensions==3.10.0.2
urllib3==1.26.7
whitenoise==5.3.0
mangum
uvloop
